feat: QeustionResponse 필드 추가

+++ src/main/java/com/example/back/question/domain/Question.java
@@ -24,7 +24,7 @@ public class Question {
 
     private double answerRate;
 
-    private short answer;
+    private int answer;
 
     private int unit;
 }
